Move cose and attestation from internal into pkg since they are used directly by selfsigned checker in nitriding
